LayoutTests/loader/load-defer-expected.txt

ALERT: Test that load deferring works as intended.
ALERT: PASS: The frame loaded after loads were resumed.
